requirements.txt
anthropic==0.49.0
boto3==1.37.33
colorama==0.4.6
coolname==2.2.0
dataclasses-json==0.6.7
genson==1.3.0
google-genai==1.10.0
humanize==4.12.2
hyperopt==0.2.7
igraph==0.11.8
jsonlines==4.0.0
jsonschema==4.23.0
kaggle==1.6.3
kornia==0.8.0
librosa==0.11.0
litellm==1.65.7
markovify==0.9.4
loguru==0.7.2
lightgbm==4.6.0
matplotlib==3.10.1
nbformat==5.10.4
openai==1.72.0
optuna==4.2.1
pandas==2.2.3
rich==14.0.0
rliable==1.2.0
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
sklearn-pandas==2.2.0
streamlit==1.44.1
sympy==1.13.3
tqdm==4.67.1
xlrd==2.0.1
submitit==1.5.2
pynvml
wandb==0.19.9
websocket-client==1.8.0
transformers==4.50.1
black==25.1.0
black[jupyter]==25.1.0
bitsandbytes==0.45.5
depyf==0.18.0
fire==0.7.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
flake8==7.2.0
hiplot==0.1.33
ipykernel==6.29.5
isort==6.0.1
pre-commit==4.2.0
jupyter-console==6.6.3
jupyterlab==4.4.0
line-profiler[ipython]==4.2.0
lovely-tensors==0.1.18
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
omegaconf==2.3.0
pandas-stubs==2.2.3.250308
pur==7.3.3
pytest==8.3.5
tables==3.10.2
tabulate==0.9.0
torchfix==0.7.0
typing-extensions==4.13.2
types-psutil==7.0.0.20250401
unidecode==1.3.8
yamllint==1.37.0
plotly==6.1.2
